Python 3 work. Change authentication from using header= to auth=.
In the course of trying to solve the SSL handshake problem, I wanted to
try python 3.3+, as it allegedly handles that. To be able to do so, I
had to get a handle on the authentication part of requests.
base64.encodestring is actually base64.encodebytes, and it ONLY accepts
bytes type objects for the conversion. If you try to % it into a string,
like I was doing with the headers for my request, they get added with a
b'...' wrapped around them.

However, requests has better ways to do authentication, which I made use
of.

Also, all strings are unicode in Python3, so took out the type check for
unicode strings.

Finally, changed the shebang to /usr/bin/env python since it's more
flexible.