Change new object creation argument to XML from str.
Posting new objects expected a str. This was problematic, as we treat
all data as xml.etree.ElementTree.Element, so there was a bunch of
conversion back and forth. This was changed to support passing XML as an
Element from the get go. Then, the only thing that has to convert it to
a string is the actual post method.

Tests were updated to send Elements instead of strings.

The other advantage is that now I can implement list operations that use
different strings for querying, e.g. computers can be queried by ID,
name, UDID, serial number, or macadress.

Removed (now) unused JSSAuthenticationError exception.