Change objects to implement a __repr__ and discard pprint method.
This seems more in line with pythonic easy, interactive working.

Rename indent to _indent to indicate that it should not be called by
clients. _indent should have a deepcopy of data sent to it, so as not to
corrupt the valid xml with a bunch of \n and crap.

Since __repr__ isn't just dumping to stdout, now we need to encode the
data to utf-8.