Begin implementing request/response adapters to wrap requests and curl
Python requests is great, but to get it to work with modern Casper
versions you need TLS, and to get that, you need an updated version of
pyopenssl. This is a hurdle that is difficult to jump for admins simply
trying to start using AutoPkgr with JSSImporter. So this commit starts
to tease out the requests API that is used for python-jss into adapters
for using requests if available, but falling back to curl.