Remove unneccessary RequestsResponseAdapter, add mising Curl args
This was wrapping something with exactly the same API, thus, remove it
to be simpler.

This commit also adds the headers argument to each of the request
methods. Finally, curl's man page specifically states that the --data
argument is equivalent to --data-ascii and requires bytes; therefore we
explicitly encode at the curl command construction phase.