Reorganize class hierarchy to make sense
JSSFlatObject got merged with JSSObject, and methods were moved up or
down from JSSObject to JSSContainerObject to avoid a lot of the
checking with class attributes that was going on. For example, we were
using `can_list` to determine if you could do a get and expect to get a
list of things. Now it just assumes that's the case for
JSSContainerObject and not for JSSObject.

This involved cleaning up _new, save, and init, but it's starting to
look a lot cleaner. Now subclasses that override JSSObject can use super
in the init to do PUT, or just do it themselves.

Renamed `list_type` to `root_tag` since that's what it is and how it is
used.

CommandFlush and LogFlush are oddballs and were dropped from being
JSSObject since they don't need to be. Now they're just plain objects
with a few methods that correspond to the API endpoints.